added custom resolvers for types
