+++ codegen/resolver.go
@@ -1,0 +1,7 @@
+package codegen
+
+import (
+	"github.com/cloudquery/plugin-sdk/schema"
+)
+
+type ResolverProvider func(path string) schema.ColumnResolver
