+++ codegen/resource.go
@@ -1,0 +1,6 @@
+package codegen
+
+type ResourceDefinition struct {
+	Name  string
+	Table *TableDefinition
+}
